+++ tests/data/project6/file2.py
@@ -1,0 +1,4 @@
+# FEWTURE-BOOM: do not report me (not an orphan)
+# FEWTURE-NOBOOM: do not report me (not an orphan)
+# FEWTURE-I-AM-AN-ORPHAN: report me
+# FEWTURE-I-AM-NOT-AN-ORPHAN: do not report me (pragma)  no-check-fixmes